+++ docs/CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 ---------
 
-#### Next Release
+#### v0.10.1 (2019-09-03)
 - Fixed issue with parallel verses loading (#104)
 - Fixed scrolling position in case of following empty verses
 - Fixed issue when loading verse detail with following empty verses (#108)